refactored tests into "process" package
