+++ .gitignore
@@ -19,3 +19,5 @@ src_wsproc/
 
 # local config
 **/resources/
+
+**/threema-connector-webtest/**
