+++ threema-connector-test/src_test/threema/connector/test/process/GetReceiverInfoTest.java
@@ -1,4 +1,4 @@
-package threema.connector.test;
+package threema.connector.test.process;
 
 import org.junit.jupiter.api.Test;
