+++ threema-connector-test/src_test/threema/connector/test/process/MessageEncryptionTest.java
@@ -1,4 +1,4 @@
-package threema.connector.test;
+package threema.connector.test.process;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
