+++ threema-connector-test/src_test/threema/connector/test/process/SingleRecipientTest.java
@@ -1,4 +1,4 @@
-package threema.connector.test;
+package threema.connector.test.process;
